requirements.txt
requests
GoogleAppEngineCloudStorageClient